requirements.txt
requests
websocket-client
twine